requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
PyJWT==2.9.0
passlib==1.7.4



